refactor & make renames visible in addressbook & add source information to tokens
