+++ app/ts/background/background-startup.ts
@@ -5,7 +5,7 @@ import { retrieveWebsiteDetails, updateExtensionBadge, updateExtensionIcon } fro
 import { clearTabStates, getSimulationResults, removeTabState, setRpcConnectionStatus, updateTabState } from './storageVariables.js'
 import { setPrependTransactionsQueue } from '../simulation/services/SimulationModeEthereumClientService.js'
 import { Simulator } from '../simulation/simulator.js'
-import { TabConnection, WebsiteTabConnections } from '../utils/user-interface-types.js'
+import { TabConnection, TabState, WebsiteTabConnections } from '../utils/user-interface-types.js'
 import { EthereumBlockHeader } from '../utils/wire-types.js'
 import { EthereumClientService } from '../simulation/services/EthereumClientService.js'
 import { getActiveAddress, getSocketFromPort, sendPopupMessageToOpenWindows, websiteSocketToString } from './backgroundUtils.js'
@@ -14,7 +14,7 @@ import { refreshSimulation } from './popupMessageHandlers.js'
 import { Semaphore } from '../utils/semaphore.js'
 import { RawInterceptedRequest } from '../utils/requests.js'
 import { verifyAccess } from './accessManagement.js'
-import { InpageScriptRequest, TabState } from '../utils/interceptor-messages.js'
+import { InpageScriptRequest } from '../utils/interceptor-messages.js'
 import { replyToInterceptedRequest } from './messageSending.js'
 import { assertNever } from '../utils/typescript.js'
 import { ICON_NOT_ACTIVE } from '../utils/constants.js'
