+++ app/ts/background/providerMessageHandlers.ts
@@ -1,5 +1,5 @@
-import { ConnectedToSigner, WalletSwitchEthereumChainReply, TabState } from '../utils/interceptor-messages.js'
-import { WebsiteTabConnections } from '../utils/user-interface-types.js'
+import { ConnectedToSigner, WalletSwitchEthereumChainReply } from '../utils/interceptor-messages.js'
+import { TabState, WebsiteTabConnections } from '../utils/user-interface-types.js'
 import { EthereumAccountsReply, EthereumChainReply } from '../utils/JsonRpc-types.js'
 import { changeActiveAddressAndChainAndResetSimulation } from './background.js'
 import { getSocketFromPort, sendInternalWindowMessage, sendPopupMessageToOpenWindows } from './backgroundUtils.js'
