+++ app/ts/components/App.tsx
@@ -3,7 +3,7 @@ import { defaultAddresses } from '../background/settings.js'
 import { SimulatedAndVisualizedTransaction, SimulationAndVisualisationResults, SimulationState, TokenPriceEstimate, RpcEntry, RpcNetwork, RpcEntries, SimulationUpdatingState, SimulationResultState } from '../utils/visualizer-types.js'
 import { ChangeActiveAddress } from './pages/ChangeActiveAddress.js'
 import { Home } from './pages/Home.js'
-import { AddressInfo, AddressInfoEntry, AddressBookEntry, SignerName, AddressBookEntries, InCompleteAddressBookEntry } from '../utils/user-interface-types.js'
+import { InCompleteAddressBookEntry, RpcConnectionStatus, TabIconDetails } from '../utils/user-interface-types.js'
 import Hint from './subcomponents/Hint.js'
 import { AddNewAddress } from './pages/AddNewAddress.js'
 import { InterceptorAccessList } from './pages/InterceptorAccessList.js'
@@ -11,12 +11,16 @@ import { ethers } from 'ethers'
 import { PasteCatcher } from './subcomponents/PasteCatcher.js'
 import { truncateAddr } from '../utils/ethereum.js'
 import { DEFAULT_TAB_CONNECTION } from '../utils/constants.js'
-import { ExternalPopupMessage, TabIconDetails, UpdateHomePage, Page, WebsiteAccessArray, WebsiteIconChanged, RpcConnectionStatus, Settings } from '../utils/interceptor-messages.js'
+import { ExternalPopupMessage, UpdateHomePage, WebsiteIconChanged, Settings } from '../utils/interceptor-messages.js'
 import { version, gitCommitSha } from '../version.js'
 import { sendPopupMessageToBackgroundPage } from '../background/backgroundUtils.js'
 import { EthereumAddress } from '../utils/wire-types.js'
 import { SettingsView } from './pages/SettingsView.js'
 import { checksummedAddress } from '../utils/bigint.js'
+import { AddressInfo, AddressInfoEntry, AddressBookEntry, AddressBookEntries } from '../utils/addressBookTypes.js'
+import { WebsiteAccessArray } from '../utils/websiteAccessTypes.js'
+import { Page } from '../utils/exportedSettingsTypes.js'
+import { SignerName } from '../utils/signerTypes.js'
 
 export function App() {
 	const [appPage, setAppPage] = useState<Page>('Home')
