+++ app/ts/components/simulationExplaining/Transactions.tsx
@@ -1,7 +1,7 @@
 import { SimulatedAndVisualizedTransaction, SimulationAndVisualisationResults, TokenVisualizerResultWithMetadata, TransactionVisualizationParameters } from '../../utils/visualizer-types.js'
 import { SmallAddress } from '../subcomponents/address.js'
 import { TokenSymbol, TokenAmount, AllApproval } from '../subcomponents/coins.js'
-import { AddressBookEntry, LogAnalysisParams, RenameAddressCallBack } from '../../utils/user-interface-types.js'
+import { LogAnalysisParams, RenameAddressCallBack } from '../../utils/user-interface-types.js'
 import { QUARANTINE_CODE, QUARANTINE_CODES_DICT } from '../../simulation/protectors/quarantine-codes.js'
 import { Error as ErrorComponent } from '../subcomponents/Error.js'
 import { identifyRoutes, identifySwap, SwapVisualization } from './SwapTransactions.js'
@@ -13,6 +13,7 @@ import { EtherTransferVisualisation, SimpleTokenTransferVisualisation } from './
 import { SimpleTokenApprovalVisualisation } from './customExplainers/SimpleTokenApprovalVisualisation.js'
 import { assertNever } from '../../utils/typescript.js'
 import { CatchAllVisualizer } from './customExplainers/CatchAllVisualizer.js'
+import { AddressBookEntry } from '../../utils/addressBookTypes.js'
 
 function isPositiveEvent(visResult: TokenVisualizerResultWithMetadata, ourAddressInReferenceFrame: bigint) {
 	if (visResult.type === 'ERC20') {
