+++ app/ts/components/subcomponents/address.tsx
@@ -1,11 +1,13 @@
 import { checksummedAddress } from '../../utils/bigint.js'
 import { Blockie } from './PreactBlocky.js'
-import { AddressBookEntry, AddressInfo, RenameAddressCallBack, Website } from '../../utils/user-interface-types.js'
+import { RenameAddressCallBack } from '../../utils/user-interface-types.js'
 import { CopyToClipboard } from './CopyToClipboard.js'
 import { ApproveIcon, ArrowIcon } from '../subcomponents/icons.js'
 import { JSX } from 'preact/jsx-runtime'
 import { useSignal } from '@preact/signals'
 import { useEffect } from 'preact/hooks'
+import { AddressBookEntry, AddressInfo } from '../../utils/addressBookTypes.js'
+import { Website } from '../../utils/websiteAccessTypes.js'
 
 export function findAddressInfo(addressToFind: bigint, addressInfos: readonly AddressInfo[]) {
 	for (const info of addressInfos) {
