+++ app/ts/utils/exportedSettingsTypes.ts
@@ -1,0 +1,68 @@
+import * as funtypes from 'funtypes'
+import { RpcNetwork } from './visualizer-types.js'
+import { EthereumQuantity, OptionalEthereumAddress } from './wire-types.js'
+import { AddressInfoArray, ContactEntries } from './addressBookTypes.js'
+import { WebsiteAccessArray } from './websiteAccessTypes.js'
+
+export type Page = funtypes.Static<typeof Page>
+export const Page = funtypes.Union(
+	funtypes.Literal('Home'),
+	funtypes.Literal('AddNewAddress'),
+	funtypes.Literal('ChangeActiveAddress'),
+	funtypes.Literal('AccessList'),
+	funtypes.Literal('ModifyAddress'),
+	funtypes.Literal('Settings'),
+)
+export const pages = funtypes.Union(Page.alternatives[0], ...Page.alternatives.map(x => x))
+export type ExportedSettings = funtypes.Static<typeof ExportedSettings>
+export const ExportedSettings = funtypes.Union(
+	funtypes.ReadonlyObject({
+		name: funtypes.Literal('InterceptorSettingsAndAddressBook'),
+		version: funtypes.Literal('1.0'),
+		exportedDate: funtypes.String,
+		settings: funtypes.ReadonlyObject({
+			activeSimulationAddress: OptionalEthereumAddress,
+			activeChain: EthereumQuantity,
+			page: Page,
+			useSignersAddressAsActiveAddress: funtypes.Boolean,
+			websiteAccess: WebsiteAccessArray,
+			simulationMode: funtypes.Boolean,
+			addressInfos: AddressInfoArray,
+			contacts: funtypes.Union(funtypes.Undefined, ContactEntries),
+			useTabsInsteadOfPopup: funtypes.Boolean,
+		})
+	}),
+	funtypes.ReadonlyObject({
+		name: funtypes.Literal('InterceptorSettingsAndAddressBook'),
+		version: funtypes.Literal('1.1'),
+		exportedDate: funtypes.String,
+		settings: funtypes.ReadonlyObject({
+			activeSimulationAddress: OptionalEthereumAddress,
+			rpcNetwork: RpcNetwork,
+			page: Page,
+			useSignersAddressAsActiveAddress: funtypes.Boolean,
+			websiteAccess: WebsiteAccessArray,
+			simulationMode: funtypes.Boolean,
+			addressInfos: AddressInfoArray,
+			contacts: funtypes.Union(funtypes.Undefined, ContactEntries),
+			useTabsInsteadOfPopup: funtypes.Boolean,
+		})
+	}),
+	funtypes.ReadonlyObject({
+		name: funtypes.Literal('InterceptorSettingsAndAddressBook'),
+		version: funtypes.Literal('1.2'),
+		exportedDate: funtypes.String,
+		settings: funtypes.ReadonlyObject({
+			activeSimulationAddress: OptionalEthereumAddress,
+			rpcNetwork: RpcNetwork,
+			page: Page,
+			useSignersAddressAsActiveAddress: funtypes.Boolean,
+			websiteAccess: WebsiteAccessArray,
+			simulationMode: funtypes.Boolean,
+			addressInfos: AddressInfoArray,
+			contacts: funtypes.Union(funtypes.Undefined, ContactEntries),
+			useTabsInsteadOfPopup: funtypes.Boolean,
+			metamaskCompatibilityMode: funtypes.Boolean,
+		})
+	}),
+)
