+++ app/ts/utils/personal-message-definitions.ts
@@ -1,12 +1,14 @@
 import * as funtypes from 'funtypes'
 import { EthereumAddress, EthereumBytes32, LiteralConverterParserFactory, NonHexBigInt } from './wire-types.js'
-import { AddressBookEntry, SignerName, Website } from './user-interface-types.js'
 import { QUARANTINE_CODE } from '../simulation/protectors/quarantine-codes.js'
 import { EthereumInput } from './wire-types.js'
 import { EnrichedEIP712 } from './eip712Parsing.js'
 import { RpcNetwork } from './visualizer-types.js'
 import { InterceptedRequest } from './requests.js'
 import { OldSignTypedDataParams, PersonalSignParams, SignTypedDataParams } from './JsonRpc-types.js'
+import { AddressBookEntry } from './addressBookTypes.js'
+import { Website } from './websiteAccessTypes.js'
+import { SignerName } from './signerTypes.js'
 
 export type EIP2612Message = funtypes.Static<typeof EIP2612Message>
 export const EIP2612Message = funtypes.ReadonlyObject({
