+++ app/ts/utils/signerTypes.ts
@@ -1,0 +1,11 @@
+import * as funtypes from 'funtypes'
+
+export type SignerName = funtypes.Static<typeof SignerName>
+export const SignerName = funtypes.Union(
+	funtypes.Literal('NoSigner'),
+	funtypes.Literal('NotRecognizedSigner'),
+	funtypes.Literal('MetaMask'),
+	funtypes.Literal('Brave'),
+	funtypes.Literal('CoinbaseWallet'),
+	funtypes.Literal('NoSignerDetected'),
+)
