+++ app/ts/utils/websiteAccessTypes.ts
@@ -1,0 +1,38 @@
+import * as funtypes from 'funtypes'
+import { EthereumAddress } from './wire-types.js'
+
+export type Website = funtypes.Static<typeof Website>
+export const Website = funtypes.ReadonlyObject({
+	websiteOrigin: funtypes.String,
+	icon: funtypes.Union(funtypes.String, funtypes.Undefined),
+	title: funtypes.Union(funtypes.String, funtypes.Undefined),
+})
+
+export type WebsiteAddressAccess = funtypes.Static<typeof WebsiteAddressAccess>
+export const WebsiteAddressAccess = funtypes.ReadonlyObject({
+	address: EthereumAddress,
+	access: funtypes.Boolean,
+}).asReadonly()
+
+export type LegacyWebsiteAccess = funtypes.Static<typeof WebsiteAccess>
+export const LegacyWebsiteAccess = funtypes.ReadonlyObject({
+	origin: funtypes.String,
+	originIcon: funtypes.Union(funtypes.String, funtypes.Undefined),
+	access: funtypes.Boolean,
+	addressAccess: funtypes.Union(funtypes.ReadonlyArray(WebsiteAddressAccess), funtypes.Undefined),
+})
+export type LegacyWebsiteAccessArray = funtypes.Static<typeof LegacyWebsiteAccessArray>
+export const LegacyWebsiteAccessArray = funtypes.ReadonlyArray(LegacyWebsiteAccess)
+
+export type WebsiteAccess = funtypes.Static<typeof WebsiteAccess>
+export const WebsiteAccess = funtypes.ReadonlyObject({
+	website: Website,
+	access: funtypes.Boolean,
+	addressAccess: funtypes.Union(funtypes.ReadonlyArray(WebsiteAddressAccess), funtypes.Undefined),
+}).asReadonly()
+
+export type WebsiteAccessArray = funtypes.Static<typeof WebsiteAccessArray>
+export const WebsiteAccessArray = funtypes.ReadonlyArray(WebsiteAccess)
+
+export type WebsiteAccessArrayWithLegacy = funtypes.Static<typeof WebsiteAccessArrayWithLegacy>
+export const WebsiteAccessArrayWithLegacy = funtypes.Union(LegacyWebsiteAccessArray, WebsiteAccessArray)
